junit5: fix WorkflowDefinitionUpdateTest
When running WorkflowDefinitionUpdateTest API returned
unexpected workflow definitions.

Spring component scan picked up a workflow that
test was expecting not to be present. For some reason this didn't
happen with H2 profile.

Modify the test to use a new workflow which is not used
by any other test and that is not component scanned.